Struct cider::utils::config::ActionConfig

pub struct ActionConfig { /* private fields */ }

Contains information required to run defined Actions

Implementations

impl ActionConfig

pub fn new(
conditions: Option<Vec<Condition>>,
retries: Option<i8>,
allowed_failure: Option<bool>,
manual: Vec<Step>
) -> Self

Creates a new ActionConfig

pub fn get_conditions(&self) -> Option<Vec<Condition>>

Gets all Conditions within an ActionConfig

pub fn set_conditions(&mut self, new_conditions: Vec<Condition>)

Changes the conditions within an ActionConfig

pub fn get_retries(&self) -> &i8

Gets the retries within an ActionConfig

pub fn set_retries(&mut self, new_retries: i8)

Changes the retries of an ActionConfig

pub fn get_allowed_failure(&self) -> &bool

Returns whether or not the Action is allowed to fail.

pub fn set_allowed_failure(&mut self, new_allowed_failure: bool)

Changes the conditions within an ActionConfig

pub fn get_manual(&self) -> &Vec<Step>

Returns the Steps to be executed by the program.

pub fn set_manual(&mut self, new_manual: Vec<Step>)

Changes the execution of Steps within an ActionConfig