Struct cider::utils::config::ShareableConfiguration

pub struct ShareableConfiguration { /* private fields */ }

Contains information that can be shared between levels of a configuration

Contains information to be shared from a higher-level of configuration to a lower-level of a configuration. This contains some “aesthetic” data, like metadata and tags, but also can contain some useful information like the source directory you would like as your root, your output directory, and more.

ShareableConfigurations also contain program-critical information like the backend, image, and the language you would like to use in order to run your Action steps.

Configuration levels are as follows: | Level | Priority | |—––|–––––| | Top | 1 | | Pipeline | 2 | | Action| 3 |

Implementations

impl ShareableConfiguration

pub fn new(
metadata: Option<HashMap<String, String>>,
title: Option<String>,
tags: Option<HashMap<String, String>>,
language: String,
image: Option<String>,
backend: String,
output: String,
source: String
) -> Self

Creates a new ShareableConfiguration

Some values are completely optional, and will either be defaulted or set as None if not provided. Note that some required information is set by default in crate::utils::parsing::json_parser if it is not explicitly defaulted here. Specifically, output, and source are defaulted to ./dist/cider and ./src, respectively.

Examples:

Basic usage:

use cider::config::ShareableConfiguration;

let s = ShareableConfiguration::new(None, None, None, "Rust".to_string(), None, "bash".to_string(), "./dist/cider".to_string(), "./src".to_string());

pub fn get_metadata(&self) -> Option<HashMap<String, String>>

Returns metadata

Returns the metadata associated with a ShareableConfiguration, and logs whether the retrieval was successful or of a None type.

Warnings

Will provide the user with a warning if metadata obtained returns a None type.

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let s = json_parser::new_top_level("./cider_config.json");

let m = s.s_config.get_metadata();

pub fn set_metadata(&mut self, new_metadata: HashMap<String, String>)

Allows the metadata of a ShareableConfiguration to be changed

Examples:

use cider::parsing::json_parser;
use std::collections::HashMap;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let mut s = json_parser::new_top_level("./cider_config.json");
let mut hm = HashMap::new();
hm.insert("some metadata tag".to_string(), "some metadata data".to_string());

let m = s.s_config.set_metadata(hm.clone());

assert_eq!(s.s_config.get_metadata().unwrap(), hm);

pub fn get_title(&self) -> Option<String>

Returns the title

Returns the title associated with a ShareableConfiguration, and logs whether the retrieval was successful or of a None type.

Warnings

Will provide the user with a warning if metadata obtained returns a None type.

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let s = json_parser::new_top_level("./cider_config.json");

let m = s.s_config.get_title();

pub fn set_title(&mut self, new_title: String)

Allows the title of a ShareableConfiguration to be changed

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let mut s = json_parser::new_top_level("./cider_config.json");
let t = "Cider".to_string();

s.s_config.set_title(t.clone());

assert_eq!(s.s_config.get_title().unwrap(), t);

pub fn get_tags(&self) -> Option<HashMap<String, String>>

Returns tags

Returns the tags associated with a ShareableConfiguration, and logs whether the retrieval was successful or of a None type.

Warnings

Will provide the user with a warning if metadata obtained returns a None type.

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let s = json_parser::new_top_level("./cider_config.json");

let m = s.s_config.get_tags();

pub fn set_tags(&mut self, new_tags: HashMap<String, String>)

Allows the tags of a ShareableConfiguration to be changed

Examples:

use cider::parsing::json_parser;
use std::collections::HashMap;
//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let mut s = json_parser::new_top_level("./cider_config.json");
let mut hm = HashMap::new();
hm.insert("some tag".to_string(), "some data".to_string());

let m = s.s_config.set_tags(hm.clone());

assert_eq!(s.s_config.get_tags().unwrap(), hm);

pub fn get_language(&self) -> &str

Returns language

Returns the language associated with a ShareableConfiguration, and logs whether the retrieval was successful or of a None type.

Warnings

Will provide the user with a warning if metadata obtained returns a None type.

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let s = json_parser::new_top_level("./cider_config.json");

let m = s.s_config.get_language();
println!("{}", m);

pub fn set_language(&mut self, new_language: String)

Allows the language of a ShareableConfiguration to be changed

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let mut s = json_parser::new_top_level("./cider_config.json");
let l = "Rust".to_string();

s.s_config.set_language(l.clone());

assert_eq!(s.s_config.get_language(), l);

pub fn get_image(&self) -> Option<String>

Returns configured image

Returns the image associated with a ShareableConfiguration, and logs whether the retrieval was successful or of a None type.

Warnings

Will provide the user with a warning if image obtained returns a None type.

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let s = json_parser::new_top_level("./cider_config.json");

let m = s.s_config.get_image();

pub fn set_image(&mut self, new_image: String)

Allows the image of a ShareableConfiguration to be changed

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let mut s = json_parser::new_top_level("./cider_config.json");
let i = "rust:1.65.0".to_string();

let m = s.s_config.set_image(i.clone());

assert_eq!(s.s_config.get_image().unwrap(), i);

pub fn get_backend(&self) -> &str

Returns backend

Returns the backend associated with a ShareableConfiguration

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let s = json_parser::new_top_level("./cider_config.json");

let m = s.s_config.get_backend();

pub fn set_backend(&mut self, new_backend: String)

Allows the backend of a ShareableConfiguration to be changed

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let mut s = json_parser::new_top_level("./cider_config.json");
let b = "bash".to_string();

s.s_config.set_backend(b.clone());

assert_eq!(s.s_config.get_backend(), b);

pub fn get_output(&self) -> &str

Returns output directory

Returns the output directory associated with a ShareableConfiguration

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let s = json_parser::new_top_level("./cider_config.json");

let m = s.s_config.get_output();

pub fn set_output(&mut self, new_output: String)

Allows the output directory of a ShareableConfiguration to be changed

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let mut s = json_parser::new_top_level("./cider_config.json");
let o = "./dist/cider".to_string();

s.s_config.set_output(o.clone());

assert_eq!(s.s_config.get_output(), o);

pub fn get_source(&self) -> &str

Returns source directory

Returns the source directory associated with a ShareableConfiguration

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let s = json_parser::new_top_level("./cider_config.json");

let m = s.s_config.get_source();

pub fn set_source(&mut self, new_source: String)

Allows the source directory of a ShareableConfiguration to be changed

Examples:

use cider::parsing::json_parser;

//returns a TopLevelConfiguration, which contains a ShareableConfiguration
let mut s = json_parser::new_top_level("./cider_config.json");
let src = "./src".to_string();

s.s_config.set_source(src.clone());

//assert_eq!(s.s_config.get_source(), &src);