Module cider::utils::executor

This module executes scripts based on the configuration provided

Structs

ExecInfo

Contains data necessary to perform specific actions in a configurable manner Combines information from both crate::utils::config::ShareableConfiguration and crate::utils::config::ActionConfig See crate::utils::config for more information.

Functions

exec_actions

Small wrapper used to gather output of multiple actions and run actions programatically